Added network calls to interactors, and added tests.

+++ app/src/main/java/mobsoft/win/the/afor/bullshit/aut/bme/autonyilvantarto/Interactor/car/CarInteractor.java
@@ -1,5 +1,8 @@
 package mobsoft.win.the.afor.bullshit.aut.bme.autonyilvantarto.Interactor.car;
 
+import android.accounts.NetworkErrorException;
+
+import java.util.ArrayList;
 import java.util.List;
 
 import javax.inject.Inject;
@@ -13,6 +16,7 @@ import mobsoft.win.the.afor.bullshit.aut.bme.autonyilvantarto.Interactor.car.eve
 import mobsoft.win.the.afor.bullshit.aut.bme.autonyilvantarto.model.Car;
 import mobsoft.win.the.afor.bullshit.aut.bme.autonyilvantarto.network.api.CarApi;
 import mobsoft.win.the.afor.bullshit.aut.bme.autonyilvantarto.repository.Repository;
+import retrofit2.Response;
 
 public class CarInteractor {
 
@@ -31,6 +35,18 @@ public class CarInteractor {
 		GetCarEvent event = new GetCarEvent();
 		try {
 			Car car = repository.getCar(id);
+			if (car == null) {
+				Response<mobsoft.win.the.afor.bullshit.aut.bme.autonyilvantarto.network.model.Car> res = carApi.carIdGet(id).execute();
+				mobsoft.win.the.afor.bullshit.aut.bme.autonyilvantarto.network.model.Car carFromNetwork = res.body();
+				car = new Car();
+				car.setYear(carFromNetwork.getYear());
+				car.setType(carFromNetwork.getType());
+				car.setPlateNumber(carFromNetwork.getPlateNumber());
+				car.setBrand(carFromNetwork.getBrand());
+				car.setEngineCapacity(carFromNetwork.getEngineCapacity());
+				car.setFuel(carFromNetwork.getFuel());
+				car.setMileage(carFromNetwork.getMileage());
+			}
 			event.setCar(car);
 			bus.post(event);
 		} catch (Exception e) {
@@ -42,8 +58,27 @@ public class CarInteractor {
 	public void getCars() {
 		GetCarsEvent event = new GetCarsEvent();
 		try {
-            carApi.carGet().execute();
-            List<Car> cars = repository.getCars();
+			List<Car> cars = repository.getCars();
+			if (cars == null || cars.size() == 0) {
+				cars = new ArrayList<Car>();
+				Response<List<mobsoft.win.the.afor.bullshit.aut.bme.autonyilvantarto.network.model.Car>> response = carApi.carGet().execute();
+				if (response.isSuccess()) {
+					for (mobsoft.win.the.afor.bullshit.aut.bme.autonyilvantarto.network.model.Car carFromNetwork : response.body()) {
+						Car car = new Car();
+						car.setYear(carFromNetwork.getYear());
+						car.setType(carFromNetwork.getType());
+						car.setPlateNumber(carFromNetwork.getPlateNumber());
+						car.setBrand(carFromNetwork.getBrand());
+						car.setEngineCapacity(carFromNetwork.getEngineCapacity());
+						car.setFuel(carFromNetwork.getFuel());
+						car.setMileage(carFromNetwork.getMileage());
+						cars.add(car);
+						repository.saveCar(car);
+					}
+				} else {
+					throw new NetworkErrorException();
+				}
+			}
 			event.setCars(cars);
 			bus.post(event);
 		} catch (Exception e) {
@@ -55,7 +90,18 @@ public class CarInteractor {
 	public void saveCar(Car car) {
 		SaveCarEvent event = new SaveCarEvent();
 		try {
-			repository.saveCar(car);
+			mobsoft.win.the.afor.bullshit.aut.bme.autonyilvantarto.network.model.Car carToPost = new mobsoft.win.the.afor.bullshit.aut.bme.autonyilvantarto.network.model.Car();
+			carToPost.setYear(car.getYear());
+			carToPost.setType(car.getType());
+			carToPost.setPlateNumber(car.getPlateNumber());
+			carToPost.setBrand(car.getBrand());
+			carToPost.setEngineCapacity(car.getEngineCapacity());
+			carToPost.setFuel(car.getFuel());
+			carToPost.setMileage(car.getMileage());
+			Response res = carApi.carPost(carToPost).execute();
+			if (res.isSuccess()) {
+				repository.saveCar(car);
+			}
 			bus.post(event);
 		} catch (Exception e) {
 			event.setThrowable(e);
@@ -66,7 +112,10 @@ public class CarInteractor {
 	public void deleteCar(Car car) {
 		DeleteCarEvent event = new DeleteCarEvent();
 		try {
-			repository.deleteCar(car);
+			Response res = carApi.carIdDelete(car.getId()).execute();
+			if (res.isSuccess()) {
+				repository.deleteCar(car);
+			}
 			bus.post(event);
 		} catch (Exception e) {
 			event.setThrowable(e);
